Rename `db._variants` to `db.variants`
